[CDEC-271] Implement asset-locking by source address
Multi-signature addresses are not yet working but are scheduled to be working
before the Cardano network becomes decentralised.

For the period before decentralisation (while Cardano is centrally mined/minted)
and while we wait for multi-sig to be implemented, this ability to asset lock
source addresses provides some extra protection for large holders of Ada like
IOHK, Emurgo and the Cardano Foundation.

Once Cardano becomes decentralised this asset-lock capability will be useless
(and will be removed) since any staking node can choose to add or remove this
feature, but it is assumed that the majority of nodes will not.